refactor(orchestrator): extract StreamEventEmitter owning event emission
Move _emit (now public emit) plus per-workflow sequence state
(_sequence_counters/_sequence_locks) and the stream-chunk mapping cluster
(handle_stream_chunk, handle_tasks_event, handle_combined_stream_chunk,
_emit_agent_messages and the per-node emit_* helpers) out of
OrchestratorService into a new StreamEventEmitter. The service composes
self._events = StreamEventEmitter(repository, event_bus) and delegates all
emission to it; cleanup callbacks call self._events.forget(workflow_id).

Tests patch orchestrator._events.emit / call orchestrator._events.handle_*
and the new test_event_emitter.py asserts on the WorkflowEvents the real
EventBus broadcasts rather than on patched internals.
