+++ tests/unit/server/orchestrator/test_service_stream_mode.py
@@ -51,7 +51,7 @@ class TestStreamModeTaskEvents:
         }
 
         # Call the handler
-        await service._handle_tasks_event(uuid4(), task_event)
+        await service._events.handle_tasks_event(uuid4(), task_event)
 
         # Verify STAGE_STARTED was emitted
         mock_event_bus.emit.assert_called()
@@ -82,7 +82,7 @@ class TestStreamModeTaskEvents:
         }
 
         # Call the handler
-        await service._handle_tasks_event(uuid4(), task_result)
+        await service._events.handle_tasks_event(uuid4(), task_result)
 
         # Verify no event was emitted
         mock_event_bus.emit.assert_not_called()
@@ -108,7 +108,7 @@ class TestStreamModeTaskEvents:
 
         # Process each chunk
         for chunk in chunks:
-            await service._handle_combined_stream_chunk(uuid4(), chunk)
+            await service._events.handle_combined_stream_chunk(uuid4(), chunk)
 
         # Verify both handlers were called: tasks mode emits STAGE_STARTED,
         # updates mode emits STAGE_COMPLETED
@@ -188,7 +188,7 @@ class TestTaskStartedEvents:
             "triggers": ["approve_plan_node"],
         }
 
-        await service._handle_tasks_event(uuid4(), task_event)
+        await service._events.handle_tasks_event(uuid4(), task_event)
 
         # Verify TASK_STARTED was emitted (in addition to STAGE_STARTED)
         assert mock_event_bus.emit.call_count == 2
